fix: Move breadcrumb under shared and fix.
