+++ nx2/blocks/browse/browse.css
@@ -66,8 +66,4 @@
     font-weight: var(--s2-component-m-bold-font-weight);
     color: var(--s2-gray-900);
   }
-
-  nx-breadcrumb {
-    display: block;
-  }
 }
